Remove player animation assets and add export icon
Deleted all player animation files and related meta files from Assets/Animations and updated PlayerAnimator.asset to not wait for completion on a state transition. Added a new export icon sprite and set it as a build target icon in ProjectSettings. Updated UniversalRenderPipelineGlobalSettings.asset to reference new resource IDs and changed the bundle version to 1.0.0b.